+++ vclib-aries/src/commonMain/kotlin/at/asitplus/wallet/lib/msg/RequestPresentationAttachment.kt
@@ -1,11 +1,10 @@
 package at.asitplus.wallet.lib.msg
 
+import at.asitplus.KmmResult.Companion.wrap
 import at.asitplus.wallet.lib.aries.jsonSerializer
 import at.asitplus.wallet.lib.data.dif.PresentationDefinition
-import io.github.aakira.napier.Napier
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.encodeToString
 
 /**
@@ -24,9 +23,6 @@ data class RequestPresentationAttachment(
     companion object {
         fun deserialize(it: String) = kotlin.runCatching {
             jsonSerializer.decodeFromString<RequestPresentationAttachment>(it)
-        }.getOrElse {
-            Napier.w("deserialize failed", it)
-            null
-        }
+        }.wrap()
     }
 }