+++ vclib-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/OidcSiopVerifier.kt
@@ -387,9 +387,8 @@ class OidcSiopVerifier(
                     return AuthnResponseResult.ValidationError("response", params.state)
                         .also { Napier.w { "JWS of response not verified: ${params.response}" } }
                 }
-                AuthenticationResponseParameters.deserialize(jarmResponse.payload.decodeToString())?.let {
-                    return validateAuthnResponse(it)
-                }
+                AuthenticationResponseParameters.deserialize(jarmResponse.payload.decodeToString())
+                    .getOrNull()?.let { return validateAuthnResponse(it) }
             }
         }
         val idTokenJws = params.idToken
@@ -401,9 +400,10 @@ class OidcSiopVerifier(
         if (!verifierJwsService.verifyJwsObject(jwsSigned))
             return AuthnResponseResult.ValidationError("idToken", params.state)
                 .also { Napier.w { "JWS of idToken not verified: $idTokenJws" } }
-        val idToken = IdToken.deserialize(jwsSigned.payload.decodeToString())
-            ?: return AuthnResponseResult.ValidationError("idToken", params.state)
-                .also { Napier.w("Could not deserialize idToken: $idTokenJws") }
+        val idToken = IdToken.deserialize(jwsSigned.payload.decodeToString()).getOrElse { ex ->
+            return AuthnResponseResult.ValidationError("idToken", params.state)
+                .also { Napier.w("Could not deserialize idToken: $idTokenJws", ex) }
+        }
         if (idToken.issuer != idToken.subject)
             return AuthnResponseResult.ValidationError("iss", params.state)
                 .also { Napier.d("Wrong issuer: ${idToken.issuer}, expected: ${idToken.subject}") }