+++ vclib/src/commonTest/kotlin/at/asitplus/wallet/lib/cbor/CoseServiceTest.kt
@@ -39,8 +39,7 @@ class CoseServiceTest : FreeSpec({
         signed.payload shouldBe randomPayload
         signed.signature.shouldNotBeNull()
 
-        val parsed = CoseSigned.deserialize(signed.serialize())
-        parsed.shouldNotBeNull()
+        val parsed = CoseSigned.deserialize(signed.serialize()).getOrThrow().shouldNotBeNull()
 
         cryptoService.coseKey shouldNotBe null
         val result = verifierCoseService.verifyCose(parsed, cryptoService.coseKey).getOrThrow()
@@ -59,8 +58,7 @@ class CoseServiceTest : FreeSpec({
         signed.payload shouldBe null
         signed.signature.shouldNotBeNull()
 
-        val parsed = CoseSigned.deserialize(signed.serialize())
-        parsed.shouldNotBeNull()
+        val parsed = CoseSigned.deserialize(signed.serialize()).getOrThrow().shouldNotBeNull()
 
         cryptoService.coseKey shouldNotBe null
         val result = verifierCoseService.verifyCose(parsed, cryptoService.coseKey).getOrThrow()